refactor(macos): extract Resolvers from AppModel into AppModel+Resolvers.swift
